Batch click counts in Redis, flush to Postgres every 30s
The click counter was a synchronous UPDATE on every redirect, so 1,000 clicks
on one link meant 1,000 writes to the same row while the browser waited.

Counts now go to a Redis hash (clicks:pending) via HINCRBY, and a @Scheduled
job folds them into Postgres every 30 seconds. A cache-hit redirect touches
Redis and nothing else.

- one hash rather than INCR clicks:{code} keys, so the flush is a single
  HGETALL instead of KEYS/SCAN to find what changed
- the flush snapshots with RENAME, which is atomic; HGETALL + DEL would drop
  every increment landing between the two
- a failed flush keeps clicks:flushing for the next cycle, so failure
  double-counts rather than loses
- getAnalytics adds the un-flushed delta, so the API never shows the lag
- the shutdown flush listens for ContextClosedEvent, not @PreDestroy:
  LettuceConnectionFactory is a SmartLifecycle and is already stopped by the
  time @PreDestroy runs, which silently stranded counts in Redis

+++ src/main/java/com/example/urlshortener/dto/AnalyticsResponse.java
@@ -8,6 +8,6 @@ public record AnalyticsResponse(
         String originalUrl,
         Instant createdAt,
         Instant expiresAt,
-        Integer totalClicks,        // from Url.clickCount (PostgreSQL)
+        long totalClicks,        // from Url.clickCount (PostgreSQL)
         List<ClickEventsDTO> clicks  // from ClickEvents collection (MongoDB)
 ) {}
